Modularize gateway API router
